+++ CONTRIBUTING.md
@@ -5,14 +5,14 @@ Thank you for showing interest in contributing to rajiniPP! When contributing to
 Please note we have a code of conduct, please follow it in all your interactions with the project.
 
 ## Ground Rules
-We expect every contribution to adhere to the following set of rules to make the project easier to maintatin and understand for everyone.
+We expect every contribution to adhere to the following set of rules to make the project easier to maintain and understand for everyone.
 1. Work on your own branch/fork, the master branch is protected.
 2. Every PR is associated with an issue.
 3. Use [conventional commit messages](https://www.conventionalcommits.org/en/v1.0.0/#specification).
 4. Adhere to the linting (flake8) and formatting (black) guidelines set up using the respective config files. Pipelines will fail if you don't.
 5. Add tests for the changes you make to add a feature or fix a bug.
 6. Ensure that the tests pass. Pipelines will fail if they don't pass.
-7. Add docstring wherewever possible. We use google-style docstring.
+7. Add docstring wherever possible. We use google-style docstring.
 
 ## Pull Request Process
